* Fixed: prevent creature's auras duplication at creature respawn. * Foxed: use expected high guid for dynamic object instead explicit outdated value. * Make some compliers more happy with using high level options for warning checks.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@5043 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/dep/src/zthread/Status.h
@@ -61,7 +61,7 @@ namespace ZThread {
 
     } STATE;
 
-    Status() : _pending(INVALID), _mask(ANYTHING) { }
+    Status() : _pending((unsigned short)INVALID), _mask((unsigned short)ANYTHING) { }
   
     /**
      * Set the mask for the STATE's that next() will report.
